Move Weights to TensorUtils
