+++ Tests/TensorUtilsTests/Resources/tensor-1d-int32.safetensors
