+++ Tests/TensorUtilsTests/Resources/tensor-2d-float64.safetensors
