+++ Tests/TensorUtilsTests/Resources/tensor-3d-float32.safetensors
