+++ Tests/TensorUtilsTests/Resources/tensor-4d-float32.safetensors
